backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
python-dotenv>=1.0.0
openai>=1.0.0
python-multipart>=0.0.6
edge-tts>=6.1.0
aiofiles>=23.2.0
chromadb>=0.4.0
numpy>=1.24.0
